+++ demo/client/webpack.config.js
@@ -2,6 +2,8 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
 const path = require('path');
 const CssMinimizerPlugin = require("css-minimizer-webpack-plugin");
+const theme = 'metronic_8.1.7';
+// const theme = 'metronic_8.1.2';
 
 module.exports = {
   watchOptions: {
@@ -14,20 +16,20 @@ module.exports = {
     ]
   },
   entry: {
-    'block-ui': './src/block-ui.js',
-    datatable: './src/datatable.js',
-    dialog: './src/dialog.js',
-    'image-input': './src/image-input.js',
-    modal: './src/modal.js',
-    toast: './src/toast.js',
-    'date-range-picker': './src/date-range-picker.js',
-    'custom-validation': './src/custom-validation.js',
-    tagify: './src/tagify.js',
-    dropzone: './src/dropzone.js',
-    'select-elements': './src/select-elements.js',
-    'toggle-password-visibility': './src/toggle-password-visibility.js',
-    'rest-client': './src/rest-client.js',
-    tree: './src/tree.js',
+    'block-ui': './src/pages/block-ui.js',
+    datatable: './src/pages/datatable.js',
+    dialog: './src/pages/dialog.js',
+    'image-input': './src/pages/image-input.js',
+    modal: './src/pages/modal.js',
+    toast: './src/pages/toast.js',
+    'date-range-picker': './src/pages/date-range-picker.js',
+    'custom-validation': './src/pages/custom-validation.js',
+    tagify: './src/pages/tagify.js',
+    dropzone: './src/pages/dropzone.js',
+    'select-elements': './src/pages/select-elements.js',
+    'toggle-password-visibility': './src/pages/toggle-password-visibility.js',
+    'rest-client': './src/pages/rest-client.js',
+    tree: './src/pages/tree.js',
   },
 
   // File output settings.
@@ -43,7 +45,9 @@ module.exports = {
     new CopyPlugin({
       patterns: [
         {from: 'src/media', to: 'media'},
-        {from: 'src/theme', to: 'theme'}
+        {from: 'src/theme/metronic_8.1.7', to: 'theme'},
+        // {from: 'src/theme/metronic_8.1.2', to: 'theme'},
+        // {from: 'src/theme', to: 'theme'},
       ]
     })
   ],