+++ docs/update-docs.js
@@ -1,21 +1,32 @@
 /**
  * This is a script that re-creates the docs based on the demo.
+ * The following process is performed
+ * - Copy demo/views/*.hbs to docs/*.html.
+ * - Copy demo/public/build to docs/build.ss
+ * - Replace references to docs/index.html links with docs/*.html.
  */
 const path = require('path');
 const fs = require('fs');
 const fse = require('fs-extra');
 
 function copyHtml() {
-  const views = fs.readdirSync(`${demoDir}/views`, {withFileTypes: true})
+  const srcFiles = fs.readdirSync(`${demoDir}/views`, {withFileTypes: true})
     .filter(dirent => dirent.isFile())
     .map(({name}) => name)
     .filter(file => path.extname(file).toLowerCase() === '.hbs');
-  for (let view of views)
-    fs.copyFileSync(`${demoDir}/views/${view}`, `${docDir}/${view}`.replace('.hbs', '.html'));
+  for (let srcFile of srcFiles) {
+    const srcPath = `${demoDir}/views/${srcFile}`;
+    const dstPath = `${docDir}/${srcFile}`.replace('.hbs', '.html');
+    console.log(`Copy ${srcPath} to ${dstPath}`);
+    fs.copyFileSync(srcPath, dstPath);
+  }
 }
 
 function copyBuild() {
-  fse.copySync(`${demoDir}/public/build`, `${docDir}/build`);
+  const srcPath = `${demoDir}/public/build`;
+  const dstPath = `${docDir}/build`;
+  console.log(`Copy ${srcPath} to ${dstPath}`);
+  fse.copySync(srcPath, dstPath);
 }
 
 function replaceIndexlinks() {
@@ -24,6 +35,7 @@ function replaceIndexlinks() {
   content = content.replace(/<a\s+(href="[a-z\-]+")..*?class="preview-thumbnail..*?">/g, match => {
     return match.replace(/href="([a-z\-]+)"/, 'href="$1.html"');
   });
+  console.log(`Replace links in ${indexHtml}`);
   fs.writeFileSync(indexHtml, content, 'utf8');
 }
 
@@ -31,5 +43,5 @@ const docDir = path.dirname(__filename);
 const rootDir = path.resolve(docDir, '..').replace(/[\/\\]$/, '');
 const demoDir = `${rootDir}/demo`;
 copyHtml();
-// copyBuild();
+copyBuild();
 replaceIndexlinks();